# anchore_engine/services/policy_engine/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a policy evaluation service. It receives push-events from external systems for data updates and provides an api for requesting image policy checks"
  version: "1.0.0"
  title: "anchore_engine.services.policy_engine"
  contact:
    email: "zach@anchore.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/v1"
schemes:
- https
- http
securityDefinitions:
  anchore_basic:
    type: basic
security:
  - anchore_basic: []
paths:
...
  /users:
    get:
      x-swagger-router-controller: anchore_engine.services.policy_engine.api.controllers.synchronous_operations
      operationId: list_image_users
      summary: "List user ids known to the eval system"
      description: "Listing of the namespaces in the service"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: query
        type: "integer"
        required: false
      responses:
        200:
          description: "User id listing"
          schema:
            type: array
            items:
              type: string
  /images:
    post:
      x-swagger-router-controller: anchore_engine.services.policy_engine.api.controllers.synchronous_operations
      operationId: ingress_image
      summary: "Add a new image to the system, as an async task"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "ingress_request"
        in: "body"
        schema:
          $ref: "#/definitions/ImageIngressRequest"
        required: true
      responses:
        200:
          description: "Image accepted"
          schema:
            $ref: "#/definitions/ImageIngressResponse"
        400:
          description: "Bad request"
  /users/{user_id}/images:
    get:
      x-swagger-router-controller: anchore_engine.services.policy_engine.api.controllers.synchronous_operations
      operationId: list_user_images
      summary: "List the image ids for the specified user"
      description: "List the image summary for each image in this user namespace"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "path"
        type: string
        description: "user id string of catalog user"
        required: true
      responses:
        200:
          description: "Image summary listing"
          schema:
            type: array
            items:
              $ref: "#/definitions/Image"
        404:
          description: "User id not found in this service"
  /users/{user_id}/images/{image_id}:
    delete:
      x-swagger-router-controller: anchore_engine.services.policy_engine.api.controllers.synchronous_operations
      operationId: delete_image
      summary: "Delete the referenced image"
      description: "Deletes the image and all artifacts from the image including cve matches"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "path"
        type: string
        required: true
      - name: "image_id"
        in: path
        type: string
        required: true
      responses:
        204:
          description: "Empty response on success"
        404:
          description: "Image not found"
        500:
          description: "Internal Error"
  /users/{user_id}/images/{image_id}/vulnerabilities:
    get:
      x-swagger-router-controller: anchore_engine.services.policy_engine.api.controllers.synchronous_operations
      operationId: get_image_vulnerabilities
      summary: "Return the vulnerabilitiy report for the image"
      description: "Returns the latest vulnerability listing for the requested image"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "path"
        type: string
        required: true
      - name: "image_id"
        in: path
        type: string
        required: true
      - name: "force_refresh"
        in: "query"
        type: boolean
        required: false
      responses:
        200:
          description: "Vulnerability listing"
          schema:
            $ref: "#/definitions/ImageVulnerabilityListing"
        404:
          description: "Image not found"
        500:
          description: "Internal Error"
  /users/{user_id}/images/{image_id}/check_inline:
    post:
      x-swagger-router-controller: anchore_engine.services.policy_engine.api.controllers.synchronous_operations
      operationId: check_user_image_inline
      summary: "List the image ids for the specified user"
      description: "List the image summary for each image in this user namespace"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "path"
        type: string
        description: "user id string of catalog user"
        required: true
      - name: "image_id"
        in: path
        type: string
        required: true
      - name: "tag"
        in: query
        type: string
        required: true
      - name: "bundle"
        in: body
        schema:
          $ref: "#/definitions/PolicyBundle"
      responses:
        200:
          description: "Policy Evaluation result"
          schema:
            $ref: "#/definitions/PolicyEvaluation"
        404:
          description: "User id not found in this service"
  /notifications/feeds:
    post:
      x-swagger-router-controller: anchore_engine.services.policy_engine.api.controllers.synchronous_operations
      operationId: create_feed_update
      description: "Send the system a feed update notification"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "notification"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/FeedUpdateNotification"
      responses:
        200:
          description: "Processed successfully"
          schema:
            type: array
            items:
              description: List of userid/image_ids updated by the sync
              type: string
        500:
          description: "Internal server error processing the request. Retry expected"
        400:
          description: "Bad request, fix and resend"
definitions:
  Image:
    type: object
    properties:
      id:
        type: string
      digest:
        type: string
      user_id:
        type: string
      state:
        type: string
        description: "State of the image in the policy evaluation system"
        enum:
        - "failed"
        - "initializing"
        - "analyzing"
        - "analyzed"
      distro_namespace:
        type: string
        description: "The namespace identifier for this image for purposes of CVE matches, etc"
      created_at:
        type: string
        format: "date-time"
        description: "The timestamp on when this image record was created, not the image itself"
      last_modified:
        type: string
        format: "date-time"
        description: "Time the image record in this service was last updated"
      tags:
        type: array
        description: "List of tags currently applied to the image. Updated by new tag events. Similarly scoped by the user_id"
        items:
          type: string
...
  FeedUpdateNotification:
    type: object
    description: "A notification event to tell the service there is new feed data to ingest. This is direct data injection"
    properties:
      event_timestamp:
        type: string
        format: "date-time"
        description: "The time of the external event. Should be set to when the event occurred, to the delivery time"
      feed_name:
        type: string
      feed_group:
        type: string
      data:
        type: "array"
        items:
          type: object
          description: "Generic object holder since each feed can have different data schemas"
...
  PolicyBundle:  
    description: "Formal schema for Anchore Policy Bundle documents"
    type: object
    required:
    - "id"
    - "version"
    properties:
      id:
        type: string
        description: "Id of the bundle"
      name:
        type: string
      comment:
        type: string
      version:
        type: string
      whitelists:
        type: "array"
        items:
          $ref: "#/definitions/Whitelist"
      policies:
        type: "array"
        items:
          $ref: "#/definitions/Policy"
      mappings:
        type: "array"
        items:
          $ref: "#/definitions/MappingRule"
  WhitelistItem:
    type: object
    required:
    - "gate"
    - "trigger_id"
    properties:
      id:
        type: string
      gate:
        type: string
      trigger_id:
        type: string
  PolicyRule:
    type: object
    required:
    - "gate"
    - "trigger"
    - "action"
    properties:
      id:
        type: string
      gate:
        type: string
      trigger:
        type: string
      action:
        enum:
        - "GO"
        - "STOP"
        - "WARN"
      params:
        type: "array"
        items:
          type: object
          required:
          - "name"
          - "value"
          properties:
            name:
              type: string
            value:
              type: string
  Whitelist:
    type: object
    required:
    - "id"
    - "version"
    properties:
      id:
        type: string
      name:
        type: string
      version:
        type: string
      comment:
        type: string
      items:
        type: "array"
        items:
          $ref: "#/definitions/WhitelistItem"
  ImageRef:
    type: object
    required:
    - "type"
    - "value"
    properties:
      type:
        enum:
          - "tag"
          - "digest"
          - "id"
      value:
        type: string
  Policy:
    type: object
    required:
    - "id"
    - "version"
    properties:
      id:
        type: string
      name:
        type: string
      comment:
        type: string
      version:
        type: string
      rules:
        type: "array"
        items:
          $ref: "#/definitions/PolicyRule"
  MappingRule:
    type: object
    required:
    - "name"
    - "whitelist_ids"
    - "policy_id"
    - "registry"
    - "repository"
    - "image"
    properties:
      id:
        type: string
      name:
        type: string
      whitelist_ids:
        type: "array"
        items:
          type: string
      policy_id:
        type: string
      registry:
        type: string
      repository:
        type: string
      image:
        $ref: "#/definitions/ImageRef"
  ImageVulnerabilityListing:
    type: object
    description: "Wrapper for the vulnerability report and legacy reports"
    properties:
      user_id:
        type: string
      image_id:
        type: string
      legacy_report:
        $ref: "#/definitions/LegacyVulnerabilityReport"
  LegacyVulnerabilityReport:
    type: object
    description: "Generic object container for the legecy cve report"
    properties:
      multi:
        type: object
        description: "Container object for a multi-image query result"
        properties:
          url_column_index:
            type: integer
            description: "The index for the url column, for parsing"
          result:
            type: object
            properties:
              rowcount:
                type: integer
              colcount:
                type: integer
              header:
                type: array
                description: "List of column names in the order the values appear in the rows section"
                items:
                  type: string
              rows:
                type: array
                description: "Data rows, each row has the same length and is colcount long"
                items:
                  type: array
                  description: "A single row entry, a list of strings in the specific order of the header"
                  items:
                    type: string
          warns:
            type: array
            description: "Array of warning messages generated during the last cve scan, if any"
            items:
              type: string
  ImageIngressRequest:
    type: object
    description: "Request to add an image to the service"
    required:
    - fetch_url
    - user_id
    - image_id
    properties:
      user_id:
        type: string
      image_id:
        type: string
      fetch_url:
        type: string
        description: "The url to pull image data from, if available. If None the system will ask the catalog endpoint"
  ImageIngressResponse:
    type: object
    description: "Indicates if the operation was accepted or not"
    properties:
      status:
        type: string
        enum:
        - accepted
        - failed
        - loaded
  PolicyEvaluation:
    type: object
    description: 'A policy bundle evaluation result for a specific image, tag, policy tuple'
    required:
      - image_id
      - bundle
      - tag
      - final_action
      - matched_mapping_rule
      - result
    properties:
      user_id:
        type: string
        description: Unique identifier (UUID) for the catalog user
      image_id:
        type: string
      tag:
        type: string
      bundle:
        type: object
        description: The bundle used for evaluation
      matched_mapping_rule:
        type: object
        description: >-
          The bundle mapping rule that was evaluated to result in the evaluated
          policy and whitelists being selected
      result:
        type: object
        description: >-
          Object containing the evaluation result for the given policy and
          whitelists against the image
      created_at:
        type: integer
        description: Epoch time on server of record creation
      last_modified:
        type: integer
        description: Epoch time on server of last modification
      final_action:
        type: string
        description: The overall outcome of the evaluation. STOP|GO|WARN
      evaluation_problems:
        type: array
        description: list of error objects indicating errors encountered during evaluation execution
        items:
          $ref: "#/definitions/PolicyEvaluationProblem"
...
  PolicyEvaluationProblem:
    type: object
    description: Details for an error or warning indicating a problem during policy evaluation
    required:
    - severity
    - problem_type
    - details
    properties:
      severity:
        type: string
        description: severity string such as "warn", "error", "fatal"
      problem_type:
        type: string
        description: the type of problem encountered, such as a misconfiguration or unavailable data
      details:
        type: string
        description: Details about the problem itself and how to fix it
...